refactored my planet(s) component, fetched data and mapped via planet, added styling to the component
