+++ src/components/planets/Planet.js
@@ -1,0 +1,14 @@
+import React, { useState, useEffect } from "react";
+
+function Planet({ planet }) {
+  return (
+    <div className="card">
+      <h3>{planet.name}</h3>
+      
+      <p>Population - {planet.population}</p>
+      <p>Terrain - {planet.terrain}</p>
+    </div>
+  );
+}
+
+export default Planet;
