+++ src/components/planets/Planets.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useQuery } from "react-query";
+import Planet from "./Planet";
 
 const fetchPlanets = async () => {
   const res = await fetch("http://swapi.dev/api/planets/");
@@ -19,9 +20,11 @@ function Planets() {
 
       {status === "error" && <div>Error fetching data</div>}
 
-      {status === 'success' && (
+      {status === "success" && (
         <div>
-          {data.results.map(planet => <div>{planet.name}</div>)}
+          {data.results.map((planet) => (
+            <Planet key={planet.name} planet={planet}/>
+          ))}
         </div>
       )}
     </>
